Reimplement on top of PR #21329
